+++ inc/admin/auth.php
@@ -262,7 +262,8 @@ function wolf_core_wrong_theme() {
 		'yor',
 		'poize',
 		'mediafoundry',
-		'soundkraft'
+		'soundkraft',
+		'aurenza'
 	);
 
 	return ( ! in_array( esc_attr( sanitize_title_with_dashes( get_template() ) ), $ok, true ) );